Change the procedure to save all the dotfiles
Before: All change was made inside MydotFiles and then copy to the HOME
directory

After: All change are made inside HOME, and then copy to MydotFiles. A
trackfile file tell witch files to copy inside MydotFiles.

This change as been made to facilitate the reload of dotfiles when doing
change.